style(clippy): clear lint backlog, shrink CI -A ratchet 17 → 2
Faz 1.2: clippy backlog temizliği. CI'daki grandfather'lı `-A` muafiyet
listesi 17 lint'ten 2'ye indi; geri kalan tüm türler artık `-D warnings`
kapısıyla zorunlu.

Düzeltilen lint türleri (hepsi davranış-koruyucu, 480 test + determinizm
stress 3/3 hash eşleşmesiyle doğrulandı):
- field_reassign_with_default (22): Default::default() + alan-atama → struct
  güncelleme literal'i
- needless_range_loop (12): index döngüleri → iterator/enumerate; tek meşru
  istisna (determinizm-kritik PGS solver) solver.rs'te yerel `#[allow]` +
  gerekçeyle dokümante edildi (global -A değil)
- manual_strip (3): elle prefix kesme → strip_prefix
- if_same_then_else (5): özdeş dallar birleştirildi (|| / tek değer)
- empty_line_after_doc (2), explicit_counter_loop (1), + --fix ile
  collapsible_if/clone_on_copy/redundant_closure/question_mark/
  assign_op_pattern/unnecessary_cast
- large_enum_variant (1): WorkerMsg::Gltf büyük payload'u Box'landı

Gerçek bug düzeltmeleri:
- lines_filter_map_ok (2, gizmo-studio/build.rs): `.lines().flatten()` okuma
  hatasında sonsuz dönebiliyordu → `.map_while(Result::ok)`
- only_used_in_recursion (1): copy_dir_all'da hiç kullanılmayan `log`
  parametresi imza + 4 çağrı yerinden kaldırıldı

-A'da bilinçli kalan 2 lint mimari (too_many_arguments, type_complexity):
fonksiyon imzalarını struct'a sarmadan refactor edilemez, ayrı bir iş.

+++ crates/gizmo-core/src/world/mod.rs
@@ -1761,7 +1761,7 @@ mod tests {
         let mut entities = Vec::new();
         for i in 0..10_000 {
             let e = world.spawn();
-            world.add_component(e, CompA(i as i32));
+            world.add_component(e, CompA(i));
             if i % 2 == 0 {
                 world.add_component(e, CompB(i as f32));
             }
